Merge branch 'tida-eventinfo-fix' into '21.1-dev' -- push to 21.0

See merge request !3526

Former-commit-id: d021e491c20d8280247acb17de12c9d57e57ae05

+++ Trigger/TrigAnalysis/TrigInDetAnalysisExample/TrigInDetAnalysisExample/Analysis_Tier0.h
@@ -185,7 +185,7 @@ class Analysis_Tier0 : public TrackAnalysis {
 
   bool     m_debug;
 
-  unsigned m_eventid;
+  unsigned long long m_eventid;
 
   VtxAnalysis* m_vtxanal;
